# lib/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ValidateError": {
				"properties": {
					"message": {
						"type": "string",
						"enum": [
							"Validation failed"
						],
						"nullable": false
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-ts-api-boilerplate",
		"version": "1.0.0",
		"description": "Boilerplate",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/echo/{echo}": {
			"get": {
				"operationId": "GetEcho",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "A Simple Echo endpoint that echoes the passed in string on the path",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "echo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/ping": {
			"get": {
				"operationId": "GetPing",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "A simple Ping message to determine if the API is available and receiving requests",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pong",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/echo-body": {
			"post": {
				"operationId": "EchoBody",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/api/form": {
			"post": {
				"operationId": "UploadForm",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"didItWork": {}
									},
									"required": [
										"didItWork"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/api/healthCheck": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"otherModles": {
											"type": "string"
										},
										"memoryAvailable": {
											"type": "number",
											"format": "double"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"otherModles",
										"memoryAvailable",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"BearerAuth": [
							"some",
							"required",
							"scopes"
						]
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}